+++ perception/approximate_deduplication/index.py
@@ -1,13 +1,20 @@
 import time
 import typing
 import warnings
+from typing import TYPE_CHECKING
 
-import faiss
 import numpy as np
-import pandas as pd
 import typing_extensions
 
 import perception.hashers.tools as pht
+from perception._optional import import_optional
+
+if TYPE_CHECKING:
+    import pandas as pd
+else:
+    pd = import_optional("pandas", extra="approximate-deduplication")
+
+faiss = import_optional("faiss", extra="approximate-deduplication")
 
 
 class QueryInput(typing_extensions.TypedDict):